feat(sync-mode): added option to propagate exceptions in sync mode
Sometimes it could be useful to stop the execution flow of the
subscribers when some of them throws an Exception.

For example if we are providing a session object and we need to
do a rollback in case something was gone wrong.

Now if we activate the option `propagateExceptionsOnSyncMode`
this will be the behaviour.